10_IT_AI_Core_Layer/backend/functions/unified_intake/requirements.txt
functions-framework==3.4.0
google-auth==2.30.0
google-api-python-client==2.134.0
google-cloud-bigquery==3.25.0
google-cloud-aiplatform==1.55.0
